+++ lizard_progress/parsers/met_parser.py
@@ -145,13 +145,13 @@ class MetParser(specifics.ProgressParser):
 
         max_z1 = max_z2 = None
 
+        self.check_waternet_profile_point_types(profile)
+        self.check_two_22_codes_with_z1z2_equal(profile)
+
         if len(profile.measurements) >= 2:
+            # Checks on the leftmost and rightmost measurement
             m1 = profile.measurements[0]
             m2 = profile.measurements[-1]
-
-            if m1.z1 != m2.z1 or m1.z2 != m2.z2:
-                self.record_error_code(
-                    profile.line_number, 'MET_LEFTRIGHTEQUAL')
 
             if m1.x == m2.x and m1.y == m2.y:
                 self.record_error_code(
@@ -200,6 +200,113 @@ class MetParser(specifics.ProgressParser):
             for measurement in profile.measurements:
                 self.check_measurement(measurement, max_z1, max_z2)
 
+    def check_waternet_profile_point_types(self, profile):
+        ## For Waternet, code checks are pretty involved:
+        # Exactly one code '1', one code '2', two codes '22',
+        # one code '7', codes '5' in between 22 and 7, codes '6' in
+        # between 7 and 22.
+        def count_codes(measurements, code):
+            return len([m for m in measurements
+                        if m.profile_point_type == code])
+
+        measurements = profile.measurements
+
+        correct_so_far = True
+
+        for code, amount, error_code in (
+            ('1', 1, 'MET_ONE_1_CODE'),
+            ('2', 1, 'MET_ONE_2_CODE'),
+            ('22', 2, 'MET_TWO_22_CODES'),
+            ('7', 1, 'MET_ONE_7_CODE')):
+            found_amount = count_codes(measurements, code)
+            if found_amount != amount:
+                self.record_error_code(
+                    profile.line_number, error_code, found_amount)
+                correct_so_far = False
+
+        if not correct_so_far:
+            return
+
+        # Now check if the 1 and 2 codes are in the correct spot
+        # We know there are more than 2 measurements or the above would
+        # never have been successful.
+        if measurements[0].profile_point_type == '1':
+            # Then last must be 2
+            if measurements[-1].profile_point_type == '2':
+                # Okay
+                pass
+            else:
+                self.record_error(
+                    measurements[-1].line_number,
+                    'MET_EXPECTED_CODE_2')
+                success_so_far = False
+        else:
+            if measurements[0].profile_point_type == '2':
+                # Then last must be 1
+                if measurements.profile_point_type == '1':
+                    # Okay. Apparently they're the other way around, so we
+                    # reverse the list for the last check.
+                    measurements = list(reversed(measurements))
+                else:
+                    self.record_error(
+                        measurements[-1].line_number,
+                        'MET_EXPECTED_CODE_2')
+                    success_so_far = False
+            else:
+                self.record_error(measurements[0].line_number,
+                                  'MET_EXPECTED_CODE_1_OR_2')
+                success_so_far = False
+
+        if not success_so_far:
+            return
+
+        # We now know the codes 1, 2, 22 and 7 have the right amounts, and
+        # 1 and 2 occur at the right places. Now:
+        # - in between 1 and 22, we may only see 99
+        # - in between 22 and 7, we may only see 5
+        # - in between 7 and 22, we may only see 6
+        # - in between 22 and 2, we may only see 99
+        indices_22 = [i for i, m in enumerate(measurements)
+                      if m.profile_point_type == '22']
+        index_7 = [i for i, m in enumerate(measurements)
+                   if m.profile_point_type == '7']
+
+        if not (indices_22[0] < index_7[0] < indices_22[1]):
+            self.record_error_code(
+                measurements[index_7[0]].line_number,
+                'MET_CODE_7_IN_BETWEEN_22')
+            return
+
+        def check_codes(measurements, code):
+            for m in measurements:
+                if m.profile_point_type != code:
+                    self.record_error_code(
+                        m.line_number,
+                        'MET_WRONG_PROFILE_POINT_TYPE',
+                        code=code)
+
+        check_codes(measurements[1:indices_22[0]], '99')
+        check_codes(measurements[indices_22[0] + 1:index_7[0]], '5')
+        check_codes(measurements[index_7[0] + 1:indices_22[1]], '6')
+        check_codes(measurements[indices_22[1] + 1:-1], '99')
+
+    def check_two_22_codes_with_z1z2_equal(self, profile):
+        if len(profile.measurements) >= 2:
+            # Checks on both 22 codes.
+            ms = [m for m in profile.measurements
+                  if m.profile_point_type == '22']
+
+            if len(ms) != 2:
+                # This error is already recorded in
+                # check_waternet_profile_point_types
+                #self.record_error_code(
+                #    profile.line_number, 'MET_TWO_22_CODES', len(ms))
+                pass
+            else:
+                if ms[0].z1 != ms[1].z1 or ms[0].z2 != ms[1].z2:
+                    self.record_error_code(
+                        profile.line_number, 'MET_LEFTRIGHTEQUAL')
+
     def check_measurement(self, measurement, max_z1, max_z2):
         if measurement.z2 < measurement.z1:
             self.record_error_code(
@@ -222,7 +329,7 @@ class MetParser(specifics.ProgressParser):
                 location = models.Location.objects.create(
                     project=self.project,
                     location_code=profile.id,
-                    the_geom=Point(profile.x, profile.y))
+                    the_geom=Point(profile.start_x, profile.start_y))
         try:
             scheduled_measurement = (
                 models.ScheduledMeasurement.objects.
@@ -242,7 +349,7 @@ class MetParser(specifics.ProgressParser):
                         project=self.project,
                         contractor=self.contractor,
                         location=location,
-                        measuremet_type=self.mtype()))
+                        measurement_type=self.mtype()))
 
         return scheduled_measurement
 
@@ -257,13 +364,13 @@ class MetParser(specifics.ProgressParser):
                 0, "MET_NOMEASUREMENTTYPE",
                 "Measurement type 'dwarsprofiel' niet geconfigureerd.")
 
-    def record_error_code(self, line_number, error_code, **kwargs):
+    def record_error_code(self, line_number, error_code, *args, **kwargs):
         if error_code not in self.error_config:
             return
 
         self.record_error(
             line_number,
-            *(models.ErrorMessage.format_code(error_code, **kwargs)))
+            *(models.ErrorMessage.format_code(error_code, *args, **kwargs)))
 
     def distance(self, m1, m2):
         try: